pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-study"
version = "0.1.0"
description = "Agentic AI interview prep — runnable demos for every concept"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "tenacity>=9.0.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]

[tool.pytest.ini_options]
testpaths = ["04_resiliency/evaluation"]
pythonpath = ["."]
